// src/types.d.ts
export interface Merchant {
  _id: string;
  email?: string;
  adress?: {
    country: string;
    state: string;
    city: string;
    street: string;
    zip: string;
  };
  currency: {
    code: string;
    symbol: string;
  };
}
// ...
export interface Product {
  _id: string;
  name: string;
  description?: string;
  price: number;
  inventory: {
    managed: boolean;
    avaible?: number;
  };
  conditionals: {
    isActive: boolean;
    isInventoryManaged: boolean; // resolverdb
    isSoldOut: boolean; //resolverdb
    hasImages: boolean; //resolverdb
  };
  image?: string;
  assets?: string[];
  sales: number; //resolverdb
  variantGroups?: VariantGroup[]; //resolverdb
  variants?: Variant[]; //resolverdb
  shippingMethods: ShippingMethod[];
  categories?: Category[];
  relatedProducts?: Product[];
  discount?: Discount; //resolverdb
  createdAt: Date;
  updatedAt: Date;
}

export interface createProductInput {
  name: string;
  description?: string;
  price: number;
  inventory: {
    managed: boolean;
    avaible?: number;
  };
  conditionals: {
    isActive: boolean;
  };
  image?: string;
  assets?: string[];
  shippingMethods: string[];
  categories?: string[];
  relatedProducts?: string[];
}

// VariantGroup
export interface VariantGroup {
  _id: string;
  name: string;
  options: string[];
  assets?: string[];
  productId: string;
}

export interface createVariantGroupInput {
  name: string;
  options: string[];
  assets?: string[];
  productId: string;
}

export interface updateVariantGroupInput {
  _id: String;
  name?: string;
  options?: string[];
  assets?: string[];
}

//Variant
export interface Variant {
  _id: string;
  inventory?: number;
  price: number;
  description?: string;
  options: {
    variantGroup: string;
    option: string;
  }[];
  assets?: string[];
  productId: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface createVariantInput {
  inventory?: number;
  price?: number;
  description?: string;
  options: {
    variantGroup: string;
    option: string;
  }[];
  assets?: string[];
  productId: string;
}

export interface updateVariantInput {
  _id: string;
  inventory?: number;
  price?: number;
  description?: string;
  assets?: string[];
}

//Category
export interface Category {
  _id: string;
  parent?: Category;
  name: string;
  description?: string;
  assets?: string[];
  children?: Category[]; // resolver
  products: Product[]; // resolver
  createdAt: Date;
  updatedAt: Date;
}

export interface createCategoryInput {
  parent?: string;
  name: string;
  description?: string;
  assets?: string[];
}

export interface updateCategoryInput {
  _id: string;
  parent?: string;
  name?: string;
  description?: string;
  assets?: string[];
}

//Cart
export interface Cart {
  _id: string;
  totalItems: number;
  subtotal: number;
  total: number;
  items: CartItem[];
  discounts: {
    code: string;
    amount: number;
  }[];
  currency: {
    code: string;
    symbol: string;
  };
  createdAt:Date
  updatedAt:Date
}

export interface CartItem {
  _id: string;
  productId: string;
  name: string;
  quantity: number;
  price: number;
  subtotal: number;
  total: number;
  variant?: string;
  image?: string;
}

export interface cartItemInput {
  productId: string;
  name: string;
  quantity: number;
  price: number;
  subtotal: number;
  total: number;
  variant?: Variant;
  image?: string;
}

//Shipping Method
export interface ShippingMethod {
  _id: string;
  name: string;
  price: number;
  zones: {
    country: string;
    regions: string[];
  }[];
}

export interface shippingMethodInput {
  name: string;
  price: number;
  zones: {
    country: string;
    regions: string[];
  }[];
}

//Discount
export interface Discount {
  _id: string;
  code: string;
  type: "percentage" | "number";
  value: number;
  startsOn: Date;
  endsOn?: Date;
  isExpired: boolean; //server
  limitQuantity?: number;
  currentQuantity: number; //server
  products: Product[];
}